Fixed the issue that each run, the xmi file is loaded in the application and somehow stored there, if you change the file and run SoMoX again, it uses the data of the old xmi file, even though it doesn't exist anymore.
Problem was that the KDMReader was using a global ResourceSet which was never deleted/overwritten and everytime SoMoX was run, the new resource (xmi file) was included at the end of the list, but for the loading process it used the first xmi resource in the list (the old one).
Now every time a SoMoX Job is started, this list is overwritten by a new one, so SoMoX just uses the actual xmi file.

git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/SoMoX/Core/trunk@30046 769cc9a0-fb0c-0410-8c2f-a2305f0e6268